+++ item_specs/file_download/src/lib.rs
@@ -11,6 +11,9 @@ pub use crate::{
     file_download_state_diff_fn_spec::FileDownloadStateDiffFnSpec,
 };
 
+#[cfg(target_arch = "wasm32")]
+pub use crate::storage_form::StorageForm;
+
 mod file_download_clean_op_spec;
 mod file_download_data;
 mod file_download_ensure_op_spec;
@@ -22,6 +25,9 @@ mod file_download_state_current_fn_spec;
 mod file_download_state_desired_fn_spec;
 mod file_download_state_diff;
 mod file_download_state_diff_fn_spec;
+
+#[cfg(target_arch = "wasm32")]
+mod storage_form;
 
 /// Read up to 1 kB in memory.
 pub const IN_MEMORY_CONTENTS_MAX: u64 = 1024;