+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## unreleased
+
+* `FileDownload` item spec now supports base64 storage for WASM target.
+
+
 ## 0.0.5 (2022-12-18)
 
 * `ShCmdStateDiffFnSpec` correctly runs `state_diff_sh_cmd` for state diff. ([#57])